further refactor improvements in order to mainain previous API contract

+++ src/main/java/dev/wms/pwrapi/entity/forum/TeacherInfoDTO.java
@@ -10,5 +10,5 @@ public class TeacherInfoDTO {
     private String category;
     private String academicTitle;
     private String fullName;
-    private Double averageRating;
+    private float averageRating;
 }
